refactor(query): applied the new folder structure to the query section related components
