+++ src/components/Form/Form.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import QuerySection from '../QuerySection/QuerySection';
+import QuerySection from '@/sections/query/QuerySection';
 import UsageSection from '../UsageSection/UsageSection';
 import DownloadSection from '../DownloadSection/DownloadSection';
