+++ src/features/query/components/AdvancedSearchForm/AdvancedSearchForm.jsx
@@ -4,6 +4,7 @@ import { Tooltip } from 'flowbite-react';
 
 import FormField from './FormFields/FormField';
 import OperatorRequest from './OperatorRequest/OperatorRequest';
+
 import useStateWithCallback from '@/hooks/useStateWithCallback';
 
 function AdvancedSearchForm ({ queryInputHandler }) {
