+++ src/features/query/components/QueryInput/QueryInput.scss
