+++ src/features/query/components/ResultPreview/ResultDetailExtension.jsx
