+++ src/features/query/components/ResultPreview/ResultList.jsx
@@ -1,5 +1,6 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
+
 import ResultDetail from './ResultDetail';
 
 export default function ResultList ({
