+++ src/features/query/components/ResultPreview/ResultLoader.jsx
