+++ src/features/query/components/ResultPreview/ResultPreview.jsx
