+++ src/sections/query/QuerySection.scss
